requirements.txt
# Core GUI Framework
PyQt5>=5.15.0

# MAVLink Communication
pymavlink>=2.4.37
pyserial>=3.5

# HTTP Requests & Image Processing
requests>=2.28.0
Pillow>=8.3.0

# Data Export
simplekml>=1.3.6

# Performance & Threading
psutil>=5.8.0

# Optional: Dark theme (remove if not needed)
# qdarkstyle>=3.1.0
